requirements.txt
torch
torchvision
transformers
Pillow
tqdm
numpy
